pyproject.toml
[project]
name = "email-marketing-agent"
version = "0.1.0"
description = "Email marketing strategy recommendation agent with LLMOps monitoring"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "pydantic>=2.7.4",
    "pydantic-settings==2.1.0",
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-community>=0.2.0",
    "faiss-cpu>=1.8.0",
    "sentence-transformers==2.2.2",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "pandas==2.1.4",
    "numpy>=1.26.0,<2",
    "tiktoken>=0.7",
    "aiofiles==23.2.1",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "python-jose[cryptography]==3.3.0",
    "python-dotenv==1.0.0",
    "structlog==23.2.0",
    "black==23.12.1",
    "isort==5.13.2",
    "langfuse>=2.0,<3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'
skip-magic-trailing-comma = true
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
src_paths = ["app", "tests"]
